Add files via upload
The Stable Diffusion Model for Image Generation on CIFAR-10 Dataset project focuses on using stable diffusion models to create images. By using the CIFAR-10 dataset, which has pictures of different objects and scenes, we aim to train a model that can make realistic images for each category. This model will understand the patterns and details in the dataset, helping it to create a wide variety of lifelike images.